refactor(api): Creates a api module to assemble stuff about api.
1. Creates an api module with a validators and a serializer files.
2. Creates a validators file in mp and member modules.
3. Adds a parameter in the documentation of the api mp module.

ZEP-23
